Move the interface. Only the index caret algorithm only able to move left and right.
